PV2-3824 Relocated files in addition to add processor/mapping logic and correcting types within models

+++ src/SFA.DAS.Payments.EarningEvents.Messages.External/EarningPeriod.cs
@@ -3,7 +3,7 @@ namespace SFA.DAS.Payments.EarningEvents.Messages.External
 {
     public class EarningPeriod
     {
-        public int DeliveryPeriod { get; set; }
+        public byte DeliveryPeriod { get; set; }
         public EarningType EarningType { get; set; }
         public decimal Amount { get; set; }
         public Employer Employer { get; set; }
